astro-airflow/requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
apache-airflow-providers-apache-spark==4.1.5
apache-airflow-providers-ssh
apache-airflow-providers-amazon
python-dotenv 
boto3